cmake: Don't include asio when RETRO_NETWORKING is OFF
asio is not needed in that case and could prevent some unsupported
platforms to build the project due to some asio #ifdef checks.
Also remove the `waitForVerify` boolean that was conditionally compiled
with networking but used in the generic path. This variable is
completely unused and never set to anything else other than false.